Add files via upload
This is a real project in progress in order to study the correlation between the students' performance of the test for enrolling in a preparatory course offered by a volunteer organization and their socioeconomic data. That course has the goal of preparing students for the college entrance exam in Brazil.